moving lib funcs
