+++ app/friends/components/FriendsToggle.tsx
@@ -4,7 +4,7 @@ import { useState, useEffect } from "react";
 import Requests from "./Requests";
 import FriendsList from "./FriendsList";
 import getFriends from "@/lib/getFriends";
-import addFriend from "@/lib/addFriend";
+import addFriend from "@/app/friends/data/addFriend";
 
 const FriendsToggle = ({ getAllRequests }: { getAllRequests: Function }) => {
   const [isFriendsList, setIsFriendsList] = useState(true);
