+++ app/friends/data/addFriend.ts
@@ -1,6 +1,6 @@
 'use server'
 
-import prisma from "../app/api/_base"
+import prisma from "../../api/_base"
 import { getServerSession } from "next-auth";
 import {authOptions} from '@/utils/authOptions'
